fin carga y manejo de datos, no tira excepcion, funcionamiento en pruebas y parece correcto.
TO-DO: metodo adyacentes

+++ SisInteligentesP1/src/infNodo.java
@@ -1,9 +1,9 @@
 
 public class infNodo {
 	private long id;
-	private float lat, lon;
+	private double lat, lon;
 
-	public infNodo(long id,float lat,float lon){
+	public infNodo(long id,double lat,double lon){
 		this.id=id;
 		this.lat=lat;
 		this.lon=lon;
